% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_matrics.R
\name{evaluate_metrics}
\alias{evaluate_metrics}
\alias{eva_ME}
\alias{eva_MAE}
\alias{eva_MSE}
\alias{eva_RMSE}
\alias{eva_R2}
\alias{eva_NSE}
\alias{eva_KGE}
\title{Evaluation metrics}
\usage{
eva_ME(num_Sim, num_Obs)

eva_MAE(num_Sim, num_Obs)

eva_MSE(num_Sim, num_Obs)

eva_RMSE(num_Sim, num_Obs)

eva_R2(num_Sim, num_Obs)

eva_NSE(num_Sim, num_Obs)

eva_KGE(num_Sim, num_Obs, num_Scal = c(1, 1, 1))
}
\arguments{
\item{num_Sim, num_Obs}{A numeric vector of simulated (\eqn{S_i}) and observed (\eqn{O_i}) values. Both vectors must have the same length.}

\item{num_Scal}{A numeric vector of length 3 with scaling factors for r, alpha, and beta in the \code{eva_KGE} calculation. Default is c(1, 1, 1).}
}
\value{
A numeric value representing the calculated error metric.
}
\description{
Evatueltion matrics with several methods.
}
\details{
\subsection{Mean Error (BIAS / ME)}{

\deqn{
\mathrm{ME} = \frac{1}{N} \sum_{i=1}^N \left( S_i - O_i \right)
}
}

\subsection{Mean Absolute Error (MAE)}{

\deqn{
\mathrm{MAE} = \frac{1}{N} \sum_{i=1}^N \left| S_i - O_i \right|
}
}

\subsection{Mean Squared Error (MSE)}{

\deqn{
\mathrm{MSE} = \frac{1}{N} \sum_{i=1}^N \left( S_i - O_i \right)^2
}
}

\subsection{Root Mean Square Error (RMSE)}{

\deqn{
\mathrm{RMSE} = \sqrt{ \frac{1}{N} \sum_{i=1}^N \left( S_i - O_i \right)^2 }
}
}

\subsection{Coefficient of Determination (R2)}{

\deqn{
R^2 = \frac{\left( \sum_{i=1}^n (O_i - \bar{O})(S_i - \bar{S}) \right)^2}{\left( \sum_{i=1}^n (O_i - \bar{O})^2 \right) \cdot \left( \sum_{i=1}^n (S_i - \bar{S})^2 \right)}
}
}

\subsection{Nash-Sutcliffe Efficiency (NSE)}{

from \insertCite{NSE_1970}{HydroRUB}

\deqn{
\mathrm{NSE} = 1 - \frac{ \sum_{i=1}^N \left( S_i - O_i \right)^2 }{ \sum_{i=1}^N \left( O_i - \bar{O} \right)^2 }
}
}

\subsection{Kling-Gupta Efficiency (KGE)}{

from \insertCite{KGE_2009,KGE_Kling_2012}{HydroRUB}

\deqn{
\mathrm{KGE} = 1 - \mathrm{ED}
}

\deqn{
\mathrm{ED} = \sqrt{ \left( s[1](r-1) \right)^2 + \left( s[2](\alpha-1) \right)^2 + \left( s[3](\beta-1) \right)^2 }
}

where

\deqn{
r = \frac{\text{cov}(S, O)}{\sigma_s \sigma_o}
}

\deqn{
\alpha = \frac{\sigma_s}{\sigma_o}
}

\deqn{
\beta = \frac{\mu_s}{\mu_o}
}
}
}
\examples{
sim <- rnorm(100)
obs <- rnorm(100)
eva_ME(sim, obs)
eva_MAE(sim, obs)
eva_MSE(sim, obs)
eva_RMSE(sim, obs)
eva_R2(sim, obs)
eva_NSE(sim, obs)
eva_KGE(sim, obs)
}
\references{
\insertRef{NSE_1970}{HydroRUB}
\insertRef{KGE_2009}{HydroRUB}
\insertRef{KGE_Kling_2012}{HydroRUB}
}
