% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HydroCourse-package.R
\docType{package}
\name{HydroCourse-package}
\alias{HydroCourse-package}
\title{A Brief Introduction to HydroCourse}
\description{
This package will gather functions and demonstration data utilized in the Ruhr-Universität Bochum course on the chair of Engineering hydrology and Water management.
}
\references{
\insertAllCited{}
}
