% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HydroRUB-package.R
\docType{_PACKAGE}
\name{HydroRUB-package}
\alias{HydroRUB-package}
\title{A Brief Introduction to HydroRUB}
\description{
This package will gather functions and demonstration data utilized in the Ruhr-Universität Bochum course on the chair of Engineering hydrology and Water management.
}
\references{
\insertAllCited{}
}
