% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{model}
\alias{model}
\alias{model_linearReservoir}
\title{Hydrological models}
\usage{
model_linearReservoir(Q_In, Q_Out0 = 0, param_K = 1)
}
\arguments{
\item{Q_In}{(vector) Inflow Boundary condition (Forcing) data}

\item{Q_Out0}{(num) Innitial condition, Outflow in the first timepoint, DEFAULT in 0}

\item{param_K}{(num) Parameter K of the linear reservoir method}
}
\value{
Outflow (vector)
}
\description{
\loadmathjax
}
\section{Linear Reservoir}{
Linear Reservoir is a method and just assuming that the watershed behaves like a \strong{linear reservoir}, where the outflow is \strong{proportional} to the water storage within the reservoir.

\mjsdeqn{Q_{out} = \frac{1}{K}S(t)}

In addition to their relationship with output and storage, linear reservoir models also adhere to the \strong{continuity equation}, often referred to as the \strong{water balance equation}.

\mjsdeqn{\frac{\mathrm{d}S(t)}{\mathrm{d}t} = Q_{in} - Q_{out}}

By combining both equations, we obtain a differential equation (DGL).

\mjsdeqn{Q_{in} = Q_{out} + K\frac{\mathrm{d}Q_{out}(t)}{\mathrm{d}t}}

\mjsdeqn{Q_{out}(t)=\int_{\tau=t0}^{t}Q_{in}(\tau)\frac{1}{K}e^{-\frac{t-\tau}{K}}\mathrm{d}\tau + Q_{out}(t_0)\frac{1}{K}e^{-\frac{t-t0}{K}}}

Where:
\itemize{
\item \mjseqn{Q_{in}} is the inflow of the reservoir
\item \mjseqn{Q_{out}} is the outflow of the reservoir
\item \mjseqn{S} is the storage of the reservoir
\item \mjseqn{K} is the parameter that defines the relationship between $Q_{out}$ and $S$
}
}

