% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_water_balance.R
\name{plot_water_balance.HBV}
\alias{plot_water_balance.HBV}
\title{Water balance plot}
\usage{
plot_water_balance.HBV(df_Model_Results, if_Perc = FALSE)
}
\arguments{
\item{df_Model_Results}{(tibble / data.frame) water balance variable from HBV models}
}
\description{
Water balance plot
}
\section{HBV Results}{
Water balance of HBV in time series with a) Flow and b) water state.
}

