% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_direction.R, R/unf.R
\name{unf}
\alias{unf}
\alias{get_inflow_cells}
\alias{read_UNF}
\alias{read_UNF_scp}
\alias{write_UNF}
\title{Some tools for flow direction / path / net}
\usage{
get_inflow_cells(num_FlowDir)

read_UNF(fn_UNF)

read_UNF_scp(fn_UNF, ip_Host, str_Username, str_Password)

write_UNF(data_Export, fn_UNF)
}
\arguments{
\item{num_FlowDir}{Integer. The cell number of the next cell. The cell number must range from 1 to the length of the cells.
If the cell has no outflow, the number should be set to itself.}

\item{fn_UNF}{string, name of the UNF file}

\item{ip_Host}{string, the name of the remote host or its IP address}

\item{str_Username}{string, the name of the user on the remote machine}

\item{str_Password}{string, a password for accessing the local SSH key}

\item{data_Export}{vector, data, that would exported}
}
\value{
A list of integers representing the cells that flow into the specified cell.

vector of numric or integer by read, UNF file by write
}
\description{
This package provides some tools for handling flow direction, path, and network-related calculations.
}
\section{Functions}{

\itemize{
\item \code{get_inflow_cells}: Get all the cells which will flow into a specified cell (including the cell itself).
}
}

\section{UNF Format}{
\itemize{
\item UNF0 / UNF8: \strong{numeric} with 4 / 8 Bytes,
\item UNF1, UNF2 / UNF4: \strong{integer} with 1, 2, / 4 Bytes
}
}

\section{Read & Write}{
\itemize{
\item Read data from UNF file
\item Write data to UNF file
}
}

\examples{
num_FD <- c(4,5,2,5,9,3,7,8,8)
lst_Cells <- get_inflow_cells(num_FD)
}
