% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HydroRUB-package.R
\docType{package}
\name{HydroRUB-package}
\alias{HydroRUB}
\alias{HydroRUB-package}
\title{A Brief Introduction to HydroCourse}
\description{
This package will gather functions and demonstration data utilized in the Ruhr-Universität Bochum course on the chair of Engineering hydrology and Water management.
}
\references{
\insertAllCited{}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/HydroSimul/HydroCourse}
  \item \url{https://hydrosimul.github.io/HydroCourse/}
  \item \url{https://hydrosimul.github.io/HydroRUB/}
}

}
